remove read me and replace anonymous iterator
